Distances

Types

Functions

SeqSpace.PointCloud.Distances.euclidean — Method

euclidean(X)

Compute the pairwise distances between points X according to the Euclidean metric. Assumes X is sized $d \times N$ where $d$ and $N$ denote dimensionality and cardinality respectively.

SeqSpace.PointCloud.Distances.euclidean² — Method

euclidean²(X)

Compute the pairwise squared distances between points X according to the Euclidean metric. Assumes X is sized $d \times N$ where $d$ and $N$ denote dimensionality and cardinality respectively.

SeqSpace.PointCloud.Distances.jensen_shannon — Method

jensen_shannon(P)

Compute the pairwise distances between probability distributions P according to the Jensen-Shannon divergence. Assumes P is sized $d \times N$ where $d$ and $N$ denote dimensionality and cardinality respectively.

SeqSpace.PointCloud.Distances.kullback_liebler — Method

kullback_liebler(p, q)

Compute the pairwise distances between probability distributions p and q according to the Kullback-Liebler divergence. Assumes p and q are normalized.