Point Cloud Generation

Types

Functions

SeqSpace.Generate.sphere — Method

sphere(N; R=1)

Generate a spherical point cloud of N points with extent of radius R.

SeqSpace.Generate.swissroll — Method

swissroll(N; z₀=10, R=1/20)

Generate a point cloud of N distributed on a swiss roll manifold with unit radius and length $rac{z₀}{R}$.

SeqSpace.Generate.torus — Method

torus(N; R=2, r=1)

Generate a point cloud of N distributed on a torus, sized inner r and outer radius R respectively.