Data Input/Output

Types

SeqSpace.DataIO.PLYProp — Type

struct PLYProp
    name :: Symbol
    type :: Type
    len  :: Union{Type,Nothing}
end

Data structure that encapsulates a property of a PLY file. If len is not nothing, it is assumed to be a collection of properties, i.e. a list.

SeqSpace.DataIO.plytype — Constant

const plytype = Dict{String, Type}

Lookup table for keywords of PLY file format matched to type information.

Functions

SeqSpace.DataIO.expand_matrix — Method

expand_matrix(io::IO, dir::String)

Read a matrix in MTX format from io stream and dump as 3 text files into directory dir.

SeqSpace.DataIO.fill! — Method

fill!(prop::Array{PLYProp}, io::IO)

Read and parse a single line from io stream. Fill the interpreted property into prop.

SeqSpace.DataIO.read_matrix — Method

read_matrix(io::IO; type=Float64, named_cols=false, named_rows=false, start_cols=1, start_rows=1)

Read a white-space delimited matrix from io stream of element type type. If named_cols is true, the first line is assumed to be column labels. If named_rows is true, the first column of each row is assumed to be denote the label.

SeqSpace.DataIO.read_mtx — Method

read_mtx(io::IO)

Read a matrix from io stream assuming bytes are formatted in matrix exchance file format.

SeqSpace.DataIO.read_obj — Method

read_obj(io::IO)

Read an oriented mesh from io stream assuming bytes are formatted in OBJ file format.

SeqSpace.DataIO.read_ply — Method

read_ply(io::IO)

Read an oriented mesh from io stream assuming bytes are formatted in PLY file format.

SeqSpace.DataIO.readprops — Method

readprops(io::IO, props::Array{PLYProp,1})

Read a collection of properties from io stream into props array. Return a named tuple of parsed properties.