Differentiable Rank

Types

Functions

SeqSpace.SoftRank.isotonic — Method

isotonic(x)

Isotonically regress on data vector x. Returns the monotonically increasing fit.

SeqSpace.SoftRank.partition — Method

partition(x; ϵ=1e-9)

Compute the sizes of each partition of constant value obtained by isotonic regression of data x. ϵ denotes the tolerance for what is considered equal in floating point terms.

SeqSpace.SoftRank.projection — Method

projection(x)

Regularize the ranking algorithm by euclidean projection onto the permutehedron.

SeqSpace.SoftRank.rank — Method

rank(x)

Rank the items of vector x in ascending order.

SeqSpace.SoftRank.softrank — Method

softrank(x)

Regularization of the ranking algorithm. Scaled euclidean projection onto the permutehedron.

SeqSpace.SoftRank.∇isotonic — Method

∇isotonic(x)

Compute the gradient of isotonic regression with respect to the inputs about the solution point.