SeqSpace API

Types

SeqSpace.HyperParams — Type

mutable struct HyperParams
    dₒ :: Int
    Ws :: Array{Int,1}
    BN :: Array{Int,1}
    DO :: Array{Int,1}
    N  :: Int
    δ  :: Int
    η  :: Float64
    B  :: Int
    V  :: Int
    k  :: Int
    γₓ :: Float32
    γᵤ :: Float32
    g  :: Function
end

HyperParams is a collection of parameters that specify the network architecture and training hyperparameters of the autoencoder. dₒ is the desired output dimensionality of the encoding layer Ws is a collection of the network layer widths. The number of entries controls the depth. The decoder is mirror-symmetric. BN is the collection of layers that will be followed by batch normalization. DO is the collection of layers that will be followed by dropout. N is the number of epochs to train against δ is the number of epochs that will be sampled for logging η is the learning rate B is the batch size V is the number of points to partition for validation purposes, i.e. won't be training against k is the number of neighbors to use to estimate geodesics γₓ is the prefactor of distance soft rank loss γᵤ is the prefactor of uniform density loss g is the metric given to latent space

SeqSpace.Result — Type

struct Result
    param :: HyperParams
    loss  :: NamedTuple{(:train, :valid), Tuple{Array{Float64,1},Array{Float64,1}} }
    model
end

Store the output of a trained autoencoder. param stores the input hyperparameters used to design and train the neural network. loss traces the dynamics of the optimization found during training. model represents the learned pullback and pushforward functions.

Functions

SeqSpace.buildloss — Method

buildloss(model, D², param)

Return a loss function used to train a neural network model according to input hyperparameters param. model is a object with three fields, pullback, pushforward, and identity. pullback and pushforward refers to the encoder and decoder layers respectively, while the identity is the composition. D² is a matrix of pairwise distances that will be used as a quenched hyperparameter in the distance soft rank loss.

SeqSpace.cylinder² — Method

cylinders²(x)

Generate the matrix of pairwise distances between vectors x, assuming the points are distributed on a cylinder. x is assumed to be $d imes N$ where $d$ denotes the dimensionality of the vector and $N$ denotes the number. The first coordinate of x is assumed to be the polar coordinate.

SeqSpace.euclidean² — Function

euclidean²(x)

Generate the matrix of pairwise distances between vectors x, assuming the Euclidean metric. x is assumed to be $d imes N$ where $d$ denotes the dimensionality of the vector and $N$ denotes the number.

SeqSpace.extendfit — Method

extendfit(result::Result, input, epochs)

Retrain model within result on input data for epochs more iterations. Returns a new Result.

SeqSpace.fitmodel — Method

fitmodel(data, param; D²=nothing, chatty=true)

Train an autoencoder model, specified with param hyperparams, to fit data. data is assumed to be sized $d imes N$ where $d$ and $N$ are dimensionality and cardinality respectively. If not nothing, D² is assumed to be a precomputed distance matrix of point cloud data. If chatty is true, function will print to stdout. Returns a Result type.

SeqSpace.linearprojection — Method

linearprojection(x, d; Δ=1, Λ=nothing)

Project an empirical distance matrix x onto d top principal components. Centers the result to have zero mean. Returns the projection, as well as a function to transform a projected vector back to the embedding space. Ignores top Δ principal components. If Λ is not nothing, assumes it is a precomputed SVD decomposition.

SeqSpace.marshal — Method

marshal(r::Result)

Serialize a trained autoencoder to binary format suitable for disk storage. Store parameters of model as contiguous array

SeqSpace.unmarshal — Method

unmarshal(r::Result)

Deserialize a trained autoencoder from binary format to semantic format. Represents model as a collection of functors.